+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - `lexicmap index, lexicmap utils edit-genome-ids/genome-details`:
     - Truncate genome/sequence IDs longer than 65,535 characters.
 - `lexicmap search`:
+    - **Faster searching speed for batch queries**.
     - Added a new flag `--show-sseq-idx` to add 1-based genome chunk and subject sequence index prefixes to sseqid values.
     - Faster pseudoalignment for long queries.
 - `lexicmap util kmers`: